Release v0.23.0

+++ crates/turin-code-index/src/code_index_writer/store.rs
@@ -146,7 +146,7 @@ pub(super) async fn delete_indexed_file(
         turso::params![relative_path.to_string()],
     )
     .await?;
-    Ok(removed_chunks as u64)
+    Ok(removed_chunks)
 }
 
 pub(super) async fn load_index_summary(conn: &turso::Connection) -> Result<CodeIndexSummary> {
